cleanup unnecessary cast

+++ src/test/java/com/rossotti/basketball/client/dto/DTOMapperTest.java
@@ -32,7 +32,7 @@ public class DTOMapperTest {
 		Assert.assertEquals("detroit-pistons", game.away_team.getTeam_id());
 		Assert.assertEquals(17, game.home_period_scores[1]);
 		Assert.assertEquals("Bojan Bogdanović", game.home_stats[0].getDisplay_name());
-		Assert.assertEquals(0f, (float)game.home_stats[0].getFree_throw_percentage(), 0.0f);
+		Assert.assertEquals(0f, game.home_stats[0].getFree_throw_percentage(), 0.0f);
 		Assert.assertEquals("Zarba", game.officials[0].getLast_name());
 		Assert.assertEquals("completed", game.event_information.getStatus());
 		Assert.assertTrue(game.away_totals.getThree_point_field_goals_attempted().equals((short)24));
